+++ ReleaseNotes.md
@@ -1,5 +1,7 @@
 # Prismic .NET Standard Release Notes
 
 ## v1.4.0
-Support boolean fragments
+* Automatically add current in when HttpContext is accessible.
+* Support boolean fragments
+* Fix caching bug